+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Golang Library Suite  
 
+## [2017-11-24]  
++ **Added**  
+  - `queue` based on single linked list in `container/queue`   
+
 ## [2017-11-22]  
 + **Updated**  
   - implement `stack` based on single linked list